pyproject.toml
[project]
name = "manimations"
version = "0.1.0"
description = "API-based educational animation generator using Manim and Claude AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "manim>=0.19.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "anthropic>=0.30.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "httpx[socks]>=0.28.1",
    "openai>=1.0.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "pdfplumber>=0.10.0",
    "asyncpg>=0.29.0",
    "databases[postgresql]>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "faker>=22.0.0",
]
